修正了jwt相关功能 修复了通用client模块失效的问题 更新了一些ID的数据类型Integer->Long便于判断 将一些决策者端需要的接口路径从拦截器中排除 基本测试完成 跨域问题配置了config但是未解决

+++ nep-aqifeedback/src/main/java/com/neu/edu/dto/AqiFeedbackDTO.java
@@ -19,7 +19,7 @@ public class AqiFeedbackDTO implements Serializable {
     private static final long serialVersionUID = 1L;
 
 	@ApiModelProperty(value = "Air Quality Public Supervision Feedback Information ID")
-	private Integer afId;
+	private Long afId;
 
 	@ApiModelProperty(value = "Public Supervisor ID (i.e., phone number)")
 	private String telId;
